refactor(api-v3): remove redundant type casts

+++ source/scripting_v3/GTA.Chrono/GameClockDate.cs
@@ -429,9 +429,7 @@ namespace GTA.Chrono
         /// </returns>
         public bool TryAdd(GameClockDuration duration, out GameClockDate date)
         {
-            // dividing int64 secs by a float64 value may result in unintended rounding errors when converting secs
-            // into float64.
-            long wholeDays = (long)duration.WholeDays;
+            long wholeDays = duration.WholeDays;
             long newOrdinal = DayOfYear + wholeDays;
 
             if (newOrdinal is > 0 and <= 365)
@@ -843,8 +841,8 @@ namespace GTA.Chrono
             DivModFloor(year1, 400, out int year1Div400, out int year1Mod400);
             DivModFloor(year2, 400, out int year2Div400, out int year2Mod400);
 
-            long cycle1 = (long)Internals.YearOrdinalToDayCycle(year1Mod400, (int)_ordFlags.Ordinal);
-            long cycle2 = (long)Internals.YearOrdinalToDayCycle(year2Mod400, (int)value._ordFlags.Ordinal);
+            long cycle1 = Internals.YearOrdinalToDayCycle(year1Mod400, (int)_ordFlags.Ordinal);
+            long cycle2 = Internals.YearOrdinalToDayCycle(year2Mod400, (int)value._ordFlags.Ordinal);
             return GameClockDuration.FromDays(((long)year1Div400 - year2Div400) * 146_097 + (cycle1 - cycle2));
         }
